moved utility functions to utils.*pp
