+++ src/HMM.cpp
@@ -876,40 +876,4 @@ void SampleGermlineState(
 };
 
 
-int ScaleMatrix(Eigen::Ref<Eigen::MatrixXd> m) {
-  int n = 0;
-
-  while ((0 < m.array() && m.array() < SCALE_THRESHOLD).any()) {
-    m *= SCALE_FACTOR;
-    n += 1;
-  }
-
-  return n;
-};
-
-
-Eigen::RowVectorXi ConvertSeqToInts(const std::string& seq_str,
-                                    const std::string& alphabet) {
-  Eigen::RowVectorXi seq(seq_str.size());
-
-  for (std::size_t i = 0; i < seq_str.size(); i++) {
-    seq[i] = GetAlphabetIndex(alphabet, seq_str[i]);
-  }
-
-  return seq;
-};
-
-
-std::string ConvertIntsToSeq(const Eigen::RowVectorXi& seq,
-                             const std::string& alphabet) {
-  std::string seq_str(seq.size(), ' ');
-
-  for (std::size_t i = 0; i < seq.size(); i++) {
-    seq_str[i] = alphabet.at(seq[i]);
-  }
-
-  return seq_str;
-};
-
-
 }  // namespace linearham
